+++ src/feedback/components/FeedbackList.vue
@@ -22,7 +22,9 @@
       <FeedbackItem
         v-for="feedbackitem in feedback"
         :key="feedbackitem.id"
+        :ref="refFor(feedbackitem.id)"
         :feedback="feedbackitem"
+        :class="{ highlight: highlight === feedbackitem.id }"
       >
         {{ $d(feedbackitem.createdAt, 'dateLongWithDayName') }}
       </FeedbackItem>
@@ -59,10 +61,33 @@ export default {
       default: null,
       type: Object,
     },
+    highlight: {
+      default: null,
+      type: Number,
+    },
   },
   computed: {
     empty () {
       return !this.isPending && !this.hasAnyErrors && this.feedback.length < 1
+    },
+    highlighted () {
+      if (!this.feedback || this.highlight < 0) return
+      return this.feedback.find(i => i.id === this.highlight)
+    },
+  },
+  watch: {
+    async highlighted (entry) {
+      if (!entry) return
+
+      // wait until element is actually rendered
+      await this.$nextTick()
+      const ref = this.$refs[this.refFor(entry.id)][0]
+      ref.$el.scrollIntoView()
+    },
+  },
+  methods: {
+    refFor (id) {
+      return `feedback-${id}`
     },
   },
 }
@@ -73,4 +98,10 @@ export default {
   max-width 600px
   margin-right auto
   margin-left auto
+
+.highlight
+  border-color $secondary
+
+  /* Override !important of no-shadow class */
+  box-shadow alpha($secondary, 0.3) 0px 0px 0px 3px !important
 </style>